removed useless defiscore code
