--- shared/src/assets/icons/defiApp.tsx
@@ -1,256 +1,0 @@
-import React from "react";
-
-type SVGProps = React.SVGAttributes<SVGElement>;
-
-export const DDEXIcon: React.FC<SVGProps> = (props) => (
-  <svg viewBox="0 0 469 431" xmlns="http://www.w3.org/2000/svg" {...props}>
-    <path d="m310 0 159 215-236.14 108.75-77.86-108.25z" fill="#17b6f0" />
-    <path d="m469 215-157 216h-2l-77.14-107.25z" fill="#00dcea" />
-    <path d="m310 0-155 215.5-155-215.5z" fill="#01c9c9" />
-    <path d="m232.86 323.75 77.14 107.25h-310z" fill="#027cad" />
-    <path d="m232.86 323.75-232.86 107.25 155-215.5z" fill="#0667b4" />
-    <path d="m0 0 155 215.5-155 215.5z" fill="#00689f" />
-  </svg>
-);
-
-export const OasisIcon: React.FC<SVGProps> = (props) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 600 600" {...props}>
-    <defs>
-      <style>
-        {`
-          .cls-1{fill:url(#linear-gradient);}.cls-2{fill:#fff;}
-        `}
-      </style>
-      <linearGradient
-        id="linear-gradient"
-        x1="300"
-        y1="602"
-        x2="300"
-        y2="2"
-        gradientTransform="matrix(1, 0, 0, -1, 0, 602)"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0" stopColor="#4fa89b" />
-        <stop offset="1" stopColor="#6acebb" />
-      </linearGradient>
-    </defs>
-    <title>Asset 1</title>
-    <g id="Layer_2" data-name="Layer 2">
-      <g id="Layer_1-2" data-name="Layer 1">
-        <circle className="cls-1" cx="300" cy="300" r="300" />
-        <path
-          className="cls-2"
-          d="M152.34,390V248.8l107.08,80.59V390h27.33V323.7A15.94,15.94,0,0,0,280.4,311L150.54,213.23A16,16,0,0,0,125,226V390Zm296.14,0V248.8L341.4,329.39V390H314.07V323.7A15.94,15.94,0,0,1,320.42,311l129.86-97.73A16,16,0,0,1,475.82,226V390Z"
-        />
-      </g>
-    </g>
-  </svg>
-);
-
-export const CompoundIcon: React.FC<SVGProps> = (props) => (
-  <svg
-    version="1.1"
-    id="Layer_1"
-    xmlns="http://www.w3.org/2000/svg"
-    x="0px"
-    y="0px"
-    viewBox="0 0 2000 2000"
-    enableBackground="new 0 0 2000 2000"
-    xmlSpace="preserve"
-    {...props}
-  >
-    <style type="text/css">
-      {`
-        .st0{fill:#070A0E;}
-        .st1{fillRule:evenodd;clipRule:evenodd;fill:#00D395;}
-      `}
-    </style>
-    <path
-      className="st0"
-      d="M1000,2000c552.3,0,1000-447.7,1000-1000S1552.3,0,1000,0S0,447.7,0,1000S447.7,2000,1000,2000z"
-    />
-    <path
-      className="st1"
-      d="M577.7,1335.3c-29.9-18.3-48.2-50.8-48.2-85.8v-195.4c0-23.2,18.9-42,42.1-41.9c7.4,0,14.7,2,21.1,5.7
-	l440.9,257.1c25.8,15,41.7,42.6,41.7,72.5v202.4c0.1,27.8-22.4,50.4-50.2,50.4c-9.3,0-18.5-2.6-26.4-7.4L577.7,1335.3z
-	 M1234.9,964.4c25.8,15,41.6,42.7,41.7,72.5v410.8c0,12.1-6.5,23.3-17.1,29.2l-96.5,54.3c-1.2,0.7-2.5,1.2-3.9,1.6v-228.1
-	c0-29.5-15.5-56.9-40.9-72.1L731,1001V743.5c0-23.2,18.9-42,42.1-41.9c7.4,0,14.7,2,21.1,5.7L1234.9,964.4z M1427.9,661
-	c25.9,15,41.8,42.7,41.8,72.6v600c-0.1,12.3-6.9,23.6-17.7,29.5l-91.5,49.4V994.8c0-29.5-15.5-56.8-40.7-72L924,685.4V441.2
-	c0-7.4,2-14.7,5.6-21.1c11.7-20,37.4-26.8,57.4-15.2L1427.9,661z"
-    />
-  </svg>
-);
-
-export const DYDXIcon: React.FC<SVGProps> = (props) => (
-  <svg
-    viewBox="0 0 183 183"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <rect
-      x="1"
-      y="1"
-      width="181"
-      height="181"
-      rx="37"
-      fill="url(#paint0_linear)"
-    />
-    <path d="M115.316 43L45 141.992H66.5885L137.269 43H115.316Z" fill="white" />
-    <path
-      d="M68.8375 43L89.5265 72.1696L78.7322 88.0804L46.7992 43H68.8375Z"
-      fill="url(#paint1_linear)"
-    />
-    <path
-      d="M117.411 142L94.4737 109.737L105.268 94.2679L139 142H117.411Z"
-      fill="url(#paint2_linear)"
-    />
-    <rect
-      x="1"
-      y="1"
-      width="181"
-      height="181"
-      rx="37"
-      stroke="#2D2D3D"
-      strokeWidth="2"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear"
-        x1="147.5"
-        y1="-24.5"
-        x2="103"
-        y2="160.5"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor="#2C2C3D" />
-        <stop offset="1" stopColor="#1a1a27" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear"
-        x1="63.8901"
-        y1="49.1875"
-        x2="92.8951"
-        y2="84.8205"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor="white" />
-        <stop offset="1" stopColor="white" stopOpacity="0.55" />
-      </linearGradient>
-      <linearGradient
-        id="paint2_linear"
-        x1="122.359"
-        y1="134.487"
-        x2="86.7873"
-        y2="85.6863"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor="#6966FF" />
-        <stop offset="1" stopColor="#6966FF" stopOpacity="0.36" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-export const AAVEIcon: React.FC<SVGProps> = (props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    clipRule="evenodd"
-    fillRule="evenodd"
-    strokeLinejoin="round"
-    strokeMiterlimit="1.41421"
-    viewBox="0 0 300.00015 300.00015"
-    version="1.1"
-    id="svg13"
-    width="300.00015"
-    height="300.00015"
-    {...props}
-  >
-    <defs id="defs17" />
-    <linearGradient
-      id="a"
-      gradientTransform="matrix(-369.036,310.497,-310.497,-369.036,434.927,94.423)"
-      gradientUnits="userSpaceOnUse"
-      x1="0"
-      x2="1"
-      y1="0"
-      y2="0"
-    >
-      <stop offset="0" stopColor="#b6509e" id="stop2" />
-      <stop offset="1" stopColor="#2ebac6" id="stop4" />
-    </linearGradient>
-    <g transform="matrix(0.604839,0,0,0.604839,-1.209678,-1.209678)" id="g11">
-      <circle cx="250" cy="250" fill="url(#a)" r="248" id="circle7" />
-      <path
-        d="M 358.5,348.8 274.6,146.1 c -4.7,-10.5 -11.8,-15.6 -21,-15.6 h -7.4 c -9.3,0 -16.3,5.1 -21,15.6 l -36.5,88.3 h -27.6 c -8.2,0.1 -15,6.7 -15,15 v 0.2 c 0.1,8.2 6.8,15 15,15 h 14.8 l -34.8,84.2 c -0.6,1.9 -1,3.8 -1,5.8 0,4.7 1.5,8.4 4.1,11.3 2.6,2.9 6.4,4.3 11.1,4.3 3.1,-0.1 6.1,-1 8.6,-2.9 2.7,-1.9 4.5,-4.5 6,-7.6 l 38.4,-95.1 h 26.6 c 8.2,-0.1 15,-6.7 15,-15 v -0.4 c -0.1,-8.2 -6.8,-15 -15,-15 h -14.2 l 29.3,-72.9 79.8,198.4 c 1.5,3.1 3.3,5.8 6,7.6 2.5,1.9 5.6,2.8 8.6,2.9 4.7,0 8.4,-1.4 11.1,-4.3 2.7,-2.9 4.1,-6.6 4.1,-11.3 0,-2 -0.3,-4 -1.1,-5.8 z"
-        fill="#ffffff"
-        fillRule="nonzero"
-        id="path9"
-      />
-    </g>
-  </svg>
-);
-
-export const YearnLogo: React.FC<SVGProps> = (props) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1000 1000" {...props}>
-    <title>yearn-finance-yfi</title>
-    <g id="Layer_2" data-name="Layer 2">
-      <g id="Layer_1-2" data-name="Layer 1">
-        <path
-          id="SVGID"
-          className="background"
-          d="M500,0c276.1,0,500,223.9,500,500s-223.9,500-500,500S0,776.1,0,500,223.9,0,500,0Z"
-          fill="#006ae3"
-          fillRule="evenodd"
-        />
-        <polygon
-          id="SVGID-2"
-          className="content"
-          data-name="SVGID"
-          points="479.4 710.8 479.4 294.8 524.7 294.8 524.7 710.8 479.4 710.8"
-          fill="#fff"
-        />
-        <path
-          id="SVGID-3"
-          className="content"
-          data-name="SVGID"
-          d="M710,433.4,570.1,470.5,538.9,325.3l41.6-9.4,16.4,68.8s37.8-62-12.6-126.2c-29.7-33-43.8-34.4-77.1-39.6-29.3-4.2-97.4,5.7-117.7,85.2-8.6,51.2,1.1,89.1,67.1,138.7L452.9,498s-73.7-51.9-92.7-88.3c-14.7-28.8-39.9-85.7,5.6-164.9,24.5-39.6,72.8-77.6,157.9-73.6C566.5,173,671,225.3,654.8,347.5c-2.8,22.9-14.7,53.4-14.7,53.4l57.4-12.8L710,433.4Z"
-          fill="#fff"
-        />
-        <path
-          id="SVGID-4"
-          className="content"
-          data-name="SVGID"
-          d="M630.1,758.9c-25.6,38.9-74.9,75.6-159.8,69.4-42.8-2.9-145.8-58-126.4-179.7,3.4-22.8,16.1-52.9,16.1-52.9l-57.7,11.2L291,561.4,431.9,528l27.4,146-41.9,8.3L402.8,613s-39.4,60.9,9.2,126.5c28.8,33.8,42.8,35.5,76.1,41.6,29.1,5,97.5-3.2,119.9-82.1,9.9-50.9,1.2-89.1-63.4-140.5l5.2-55.1s72.3,53.8,90.3,90.7c13.8,29.3,37.5,86.9-10,164.8Z"
-          fill="#fff"
-        />
-      </g>
-    </g>
-  </svg>
-);
-
-export const GnosisLogo: React.FC<SVGProps> = (props) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 2000 2000" {...props}>
-    <g fill="#00a6c4">
-      <path d="M1809.4 500l-73.6 73.6c58.5 92 76.9 207.4 36.8 317.7-66.9 187.3-274.2 286-463.2 219.1-26.8-10-51.8-21.7-73.6-36.8L1000 1309.4 794.3 1102c-95.3 63.5-217.4 81.9-332.8 41.8-194-70.2-294.3-282.6-224.1-474.9 10-30.1 25.1-56.9 41.8-81.9l-88.6-88.6-16.7 28.4c-92 150.5-142.1 324.4-142.1 503.3-1.7 533.4 433.1 969.9 966.5 969.9h1.7c533.4 0 966.6-433.1 968.2-966.6 0-177.3-48.5-351.2-140.5-503.3l-18.3-30.1" />
-      <path d="M388 695.7c-25.1 33.4-40.1 75.3-40.1 120.4 0 108.7 88.6 197.3 197.3 197.3 45.2 0 87-15.1 120.4-41.8L388 695.7M1356.2 954.8c31.8 21.7 68.6 33.4 110.4 33.4 108.7 0 197.3-88.6 197.3-197.3 0-40.1-11.7-78.6-33.4-110.4l-274.3 274.3M1003.3 1162.2L204 359.5l31.8-33.4C433.1 117.1 702.3 0 991.6 0h1.7c292.6 0 573.6 125.4 769.2 342.8l30.1 33.4-789.3 786M334.5 359.5l668.9 670.6 660.5-657.2C1486.6 195.7 1245.8 93.6 995 93.6h-1.7c-249.2 0-481.6 93.7-658.8 265.9" />
-    </g>
-  </svg>
-);
-
-export const FlexLogo: React.FC<SVGProps> = (props) => (
-  <svg
-    width="27"
-    height="24"
-    viewBox="0 0 27 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path d="M0 24V0H3.15328V8.05839H10.6861L0 24Z" fill="#14F195" />
-    <path
-      d="M26.2774 24H22.9489V16.1168H15.2409L26.2774 0V24Z"
-      fill="#14F195"
-    />
-  </svg>
-);
