+++ shared/src/assets/icons/icons.tsx
@@ -134,44 +134,6 @@ export const BarChartIcon: React.FC<IconProps> = ({
   </span>
 );
 
-export const GlobeIcon: React.FC<IconProps> = ({
-  containerStyle = { display: "flex" },
-  color = "white",
-  onClick = () => {},
-  ...props
-}) => (
-  <span style={containerStyle}>
-    <svg
-      width="20"
-      height="20"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      onClick={onClick}
-      {...props}
-    >
-      <path
-        d="M10 18.3333C14.6024 18.3333 18.3334 14.6024 18.3334 10C18.3334 5.39763 14.6024 1.66667 10 1.66667C5.39765 1.66667 1.66669 5.39763 1.66669 10C1.66669 14.6024 5.39765 18.3333 10 18.3333Z"
-        stroke={color}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M1.66669 10H18.3334"
-        stroke={color}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M10 1.66667C12.0844 3.94863 13.269 6.91003 13.3334 10C13.269 13.09 12.0844 16.0514 10 18.3333C7.91562 16.0514 6.73106 13.09 6.66669 10C6.73106 6.91003 7.91562 3.94863 10 1.66667V1.66667Z"
-        stroke={color}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  </span>
-);
-
 export const GridIcon: React.FC<SVGProps> = (props) => (
   <svg
     width="20"
