+++ shared/src/components/Product/Theta/YieldFrame.tsx
@@ -1,8 +1,7 @@
 import { formatUnits } from "@ethersproject/units";
 import { AnimatePresence, motion } from "framer";
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import styled from "styled-components";
-import { BarChartIcon, GlobeIcon } from "../../../assets/icons/icons";
 
 import {
   isPutVault,
@@ -16,7 +15,6 @@ import {
   Title,
 } from "../../../designSystem";
 import theme from "../../../designSystem/theme";
-import useAssetsYield from "../../../hooks/useAssetsYield";
 import useLatestAPY from "../../../hooks/useLatestAPY";
 import useLoadingText from "../../../hooks/useLoadingText";
 import { useV2VaultData, useVaultData } from "../../../hooks/web3DataContext";
@@ -24,13 +22,12 @@ import { getAssetLogo } from "../../../utils/asset";
 import { formatSignificantDecimals } from "../../../utils/math";
 import { getVaultColor } from "../../../utils/vault";
 import CapBar from "../../Deposit/CapBar";
-import YieldComparison from "./YieldComparison";
 
 const FrameContainer = styled.div`
   perspective: 2000px;
 `;
 
-const Frame = styled(motion.div)<{ color: string }>`
+const Frame = styled(motion.div) <{ color: string }>`
   display: flex;
   flex-direction: column;
   flex-wrap: wrap;
@@ -58,21 +55,10 @@ const TagContainer = styled.div`
   margin-right: auto;
 `;
 
-const ProductTag = styled(BaseButton)<{ color: string }>`
+const ProductTag = styled(BaseButton) <{ color: string }>`
   background: ${(props) => props.color}29;
   padding: 8px;
   margin-right: 4px;
-`;
-
-const ModeSwitcherContainer = styled.div<{ color: string }>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 100px;
-  background: ${(props) => props.color}14;
-  z-index: 1;
 `;
 
 interface YieldFrameProps {
@@ -94,7 +80,6 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
     data: { totalBalance: v2Deposits, cap: v2VaultLimit },
     loading: v2DataLoading,
   } = useV2VaultData(vault);
-  const yieldInfos = useAssetsYield(asset);
   const isLoading = useMemo(
     () => status === "loading" || v2DataLoading,
     [status, v2DataLoading]
@@ -106,8 +91,6 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
     : loadingText;
   const color = getVaultColor(vault);
   const Logo = getAssetLogo(displayAsset);
-
-  const [mode, setMode] = useState<"info" | "yield">("info");
 
   const [totalDepositStr, depositLimitStr] = useMemo(() => {
     switch (vaultVersion) {
@@ -132,11 +115,6 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
     }
   }, [decimals, deposits, v2Deposits, v2VaultLimit, vaultLimit, vaultVersion]);
 
-  const onSwapMode = useCallback((e) => {
-    e.stopPropagation();
-    setMode((prev) => (prev === "info" ? "yield" : "info"));
-  }, []);
-
   const logo = useMemo(() => {
     switch (displayAsset) {
       case "yvUSDC":
@@ -155,69 +133,39 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
   }, [displayAsset, Logo]);
 
   const body = useMemo(() => {
-    switch (mode) {
-      // @ts-ignore
-      case "yield":
-        if (yieldInfos) {
-          return (
-            <YieldComparison
-              vault={vault}
-              vaultVersion={vaultVersion}
-              config={{
-                background: `${color}29`,
-              }}
-              yieldInfos={yieldInfos}
-            />
-          );
-        }
-      // eslint-disable-next-line no-fallthrough
-      default:
-        return (
-          <>
-            <div className="mt-4 d-flex justify-content-center">{logo}</div>
-            <div className="d-flex align-items-center mt-4">
-              <SecondaryText fontSize={12} lineHeight={16} className="mr-auto">
-                Projected Yield (APY)
-              </SecondaryText>
-              <Title fontSize={14} lineHeight={20} color={color}>
-                {perfStr}
-              </Title>
-            </div>
-            <div className="mt-auto">
-              <CapBar
-                loading={isLoading}
-                current={totalDepositStr}
-                cap={depositLimitStr}
-                copies={{
-                  current: "Current Deposits",
-                  cap: "Max Capacity",
-                }}
-                labelConfig={{
-                  fontSize: 12,
-                }}
-                statsConfig={{
-                  fontSize: 14,
-                }}
-                barConfig={{ height: 4, extraClassNames: "my-2", radius: 2 }}
-                asset={asset}
-              />
-            </div>
-          </>
-        );
-    }
-  }, [
-    vault,
-    vaultVersion,
-    logo,
-    asset,
-    mode,
-    color,
-    perfStr,
-    isLoading,
-    totalDepositStr,
-    depositLimitStr,
-    yieldInfos,
-  ]);
+    return (
+      <>
+        <div className="mt-4 d-flex justify-content-center">{logo}</div>
+        <div className="d-flex align-items-center mt-4">
+          <SecondaryText fontSize={12} lineHeight={16} className="mr-auto">
+            Projected Yield (APY)
+          </SecondaryText>
+          <Title fontSize={14} lineHeight={20} color={color}>
+            {perfStr}
+          </Title>
+        </div>
+        <div className="mt-auto">
+          <CapBar
+            loading={isLoading}
+            current={totalDepositStr}
+            cap={depositLimitStr}
+            copies={{
+              current: "Current Deposits",
+              cap: "Max Capacity",
+            }}
+            labelConfig={{
+              fontSize: 12,
+            }}
+            statsConfig={{
+              fontSize: 14,
+            }}
+            barConfig={{ height: 4, extraClassNames: "my-2", radius: 2 }}
+            asset={asset}
+          />
+        </div>
+      </>
+    );
+  }, [logo, asset, color, perfStr, isLoading, totalDepositStr, depositLimitStr]);
 
   return (
     <FrameContainer
@@ -226,7 +174,6 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
     >
       <AnimatePresence exitBeforeEnter initial={false}>
         <Frame
-          key={mode}
           transition={{
             duration: 0.1,
             type: "keyframes",
@@ -252,21 +199,6 @@ const YieldFrame: React.FC<YieldFrameProps> = ({
                 </Subtitle>
               </ProductTag>
             </TagContainer>
-
-            {/* Mode switcher button */}
-            {Boolean(yieldInfos && yieldInfos.length) && (
-              <ModeSwitcherContainer
-                role="button"
-                onClick={onSwapMode}
-                color={color}
-              >
-                {mode === "info" ? (
-                  <GlobeIcon color={color} />
-                ) : (
-                  <BarChartIcon color={color} />
-                )}
-              </ModeSwitcherContainer>
-            )}
           </div>
           {body}
         </Frame>
